% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_topology.r
\name{catch2graph}
\alias{catch2graph}
\title{Calculate graph object based on catchment network}
\usage{
catch2graph(catch)
}
\arguments{
\item{catch}{a sf dataframe from HydroSheds.}
}
\value{
g a igraph object
}
\description{
Calculate graph object based on catchment network
}
