% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{catchment_geometry}
\alias{catchment_geometry}
\title{Catchments of Otto-level 12 for Ceará}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 1235 rows and 14 columns.}
\source{
https://hydrosheds.org/downloads
}
\usage{
data(catchment_geometry)
}
\description{
Catchments of Otto-level 12 for Ceará
}
\references{
Lehner, B., Verdin, K., Jarvis, A. (2008): New global hydrography derived from spaceborne elevation data. Eos, Transactions, AGU, 89(10): 93-94.
}
\keyword{datasets}
