% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{flow_direction_tidygraph}
\alias{flow_direction_tidygraph}
\title{A graph of the class \code{tidygraph} containing flow direction derived from the HydroSHEDS hydrologically conditioned DEM in Ceará with a resolution of 30 arcsec.}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\source{
https://hydrosheds.org/downloads
}
\usage{
data(flow_direction_tidygraph)
}
\description{
A graph of the class \code{tidygraph} containing flow direction derived from the HydroSHEDS hydrologically conditioned DEM in Ceará with a resolution of 30 arcsec.
}
\references{
Lehner, B., Verdin, K., Jarvis, A. (2008): New global hydrography derived from spaceborne elevation data. Eos, Transactions, AGU, 89(10): 93-94.
}
\keyword{datasets}
