% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{flow_graph_in_polygon}
\alias{flow_graph_in_polygon}
\title{Obtain flow graph inside reservoir polygon based on hydrosheds condem direction raster}
\usage{
flow_graph_in_polygon(sf_polygon)
}
\arguments{
\item{sf_polygon}{is the reservoir polygon in the same crs as the rasters \code{raster_flows_from} and \code{raster_flows_to}}
}
\value{
flow graph is a tidygraph where each node is a named pixel of the hydrosheds condem inside the input sf_polygon
}
\description{
Obtain flow graph inside reservoir polygon based on hydrosheds condem direction raster
}
