% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{graph_reverse_direction}
\alias{graph_reverse_direction}
\title{Reverse the edge direction of a directed graph}
\usage{
graph_reverse_direction(graph)
}
\arguments{
\item{graph}{a igraph directed graph}
}
\value{
reversed a graph with reversed direction
}
\description{
Reverse the edge direction of a directed graph
}
