% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Nodes of the river network with upstream cell number for Ceará. Can be computed with \code{riv2nodes}}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3773 rows and 3 columns.
}
\source{
https://hydrosheds.org/downloads
}
\usage{
data(nodes)
}
\description{
Nodes of the river network with upstream cell number for Ceará. Can be computed with \code{riv2nodes}
}
\references{
Lehner, B., Verdin, K., Jarvis, A. (2008): New global hydrography derived from spaceborne elevation data. Eos, Transactions, AGU, 89(10): 93-94.
}
\keyword{datasets}
