% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{outlet_in_condem}
\alias{outlet_in_condem}
\title{Get outlet of reservoir in hydrosheds condem raster}
\usage{
outlet_in_condem(res_id = 34181)
}
\arguments{
\item{id}{is the id of the reservoir as given in reservoir_geometry_raw in column id_jrc}
}
\value{
pixel_id that can be matched with flows_from.tiff or with flow_direction_tidygraph
}
\description{
Get outlet of reservoir in hydrosheds condem raster
}
