% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{raster_flows_to}
\alias{raster_flows_to}
\title{A raster of the class \code{raster} where each cell contains the id
of the destination cell that can be obtained in \code{flows_from}.
This raster was derived from the HydroSHEDS hydrologically
conditioned DEM in Ceará with a resolution of 30 arcsec.}
\format{
An object of class \code{RasterLayer} of dimension 658 x 729 x 1.
}
\source{
https://hydrosheds.org/downloads
}
\usage{
data(raster_flows_to)
}
\description{
A raster of the class \code{raster} where each cell contains the id
of the destination cell that can be obtained in \code{flows_from}.
This raster was derived from the HydroSHEDS hydrologically
conditioned DEM in Ceará with a resolution of 30 arcsec.
}
\references{
Lehner, B., Verdin, K., Jarvis, A. (2008): New global hydrography derived from spaceborne elevation data. Eos, Transactions, AGU, 89(10): 93-94.
}
\keyword{datasets}
