% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{reservoir_geometry}
\alias{reservoir_geometry}
\title{Geometry of reservoirs in Ceará}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 22911 rows and 13 columns.}
\source{
https://global-surface-water.appspot.com/download
}
\usage{
data(reservoir_geometry)
}
\description{
Geometry of reservoirs in Ceará
}
\references{
Jean-Francois Pekel, Andrew Cottam, Noel Gorelick, Alan S. Belward, High-resolution mapping of global surface water and its long-term changes. Nature 540, 418-422 (2016). (doi:10.1038/nature20584)
}
\keyword{datasets}
