% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{reservoir_geometry_raw}
\alias{reservoir_geometry_raw}
\title{Geometry of reservoirs in Ceará as imported from jrc}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 22911 rows and 4 columns.
}
\source{
https://global-surface-water.appspot.com/download
}
\usage{
data(reservoir_geometry_raw)
}
\description{
Geometry of reservoirs in Ceará as imported from jrc
}
\references{
Jean-Francois Pekel, Andrew Cottam, Noel Gorelick, Alan S. Belward, High-resolution mapping of global surface water and its long-term changes. Nature 540, 418-422 (2016). (doi:10.1038/nature20584)
}
\keyword{datasets}
