% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{reservoir_graph}
\alias{reservoir_graph}
\title{A graph of the class \code{igraph}: routing of strategic and non-strategic reservoirs in Ceará.}
\format{An object of class \code{igraph} of length 10.}
\source{
https://hydrosheds.org/downloads

https://global-surface-water.appspot.com/download
}
\usage{
data(reservoir_graph)
}
\description{
A graph of the class \code{igraph}: routing of strategic and non-strategic reservoirs in Ceará.
}
\keyword{datasets}
