% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{reservoir_tidygraph}
\alias{reservoir_tidygraph}
\title{A graph of the class \code{tidygraph}. Good for manipulating attributes of reservoirs as a graph.}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\source{
https://hydrosheds.org/downloads

https://global-surface-water.appspot.com/download
}
\usage{
data(reservoir_tidygraph)
}
\description{
A graph of the class \code{tidygraph}. Good for manipulating attributes of reservoirs as a graph.
}
\keyword{datasets}
