% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{riv2graph}
\alias{riv2graph}
\title{Calculate graph object based on river network}
\usage{
riv2graph(nodes_i, riv_i)
}
\arguments{
\item{nodes_i}{a sf dataframe of points marking the nodes of the river network defined as the inlet of each river reach. This must be only one tree. It won't work with a forest.}

\item{riv_i}{a sf dataframe with a topologicaly valid river network. This must be only one tree. It won't work with a forest.}
}
\value{
g a igraph object
}
\description{
Calculate graph object based on river network
}
