% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{riv2nodes}
\alias{riv2nodes}
\title{Obtain graph nodes from HydroSheds river network}
\usage{
riv2nodes(riv)
}
\arguments{
\item{riv}{a sf dataframe with a topologicaly valid river network. Each linestrings is ordered from upstream (first point of the linestring) to downstream (last point of the linestring), just like in the HydroSheds dataset}
}
\value{
nodes a sf dataframe of points marking the nodes of the river network defined as the inlet of each river reach.
}
\description{
Extract the nodes of the HydroSheds river network.
}
