% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{river_geometry}
\alias{river_geometry}
\title{River network with upstream cell number for Ceará}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7034 rows and 6 columns.
}
\source{
https://hydrosheds.org/downloads
}
\usage{
data(river_geometry)
}
\description{
River network with upstream cell number for Ceará
}
\references{
Lehner, B., Verdin, K., Jarvis, A. (2008): New global hydrography derived from spaceborne elevation data. Eos, Transactions, AGU, 89(10): 93-94.
}
\keyword{datasets}
