% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5udes-data.R
\docType{data}
\name{river_graph}
\alias{river_graph}
\title{A graph of the class \code{igraph} containing flow direction for the HydroSHEDS river network dataset in Ceará.}
\format{An object of class \code{igraph} of length 10.}
\source{
https://hydrosheds.org/downloads
}
\usage{
data(river_graph)
}
\description{
A graph of the class \code{igraph} containing flow direction for the HydroSHEDS river network dataset in Ceará.
}
\references{
Lehner, B., Verdin, K., Jarvis, A. (2008): New global hydrography derived from spaceborne elevation data. Eos, Transactions, AGU, 89(10): 93-94.
}
\keyword{datasets}
