% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{route_reservoir_along_river}
\alias{route_reservoir_along_river}
\title{Routing reservoirs across the river network}
\usage{
route_reservoir_along_river(reservoir_geometry_subset)
}
\arguments{
\item{reservoir_geometry_subset}{is  a subset \code{data(reservoir_geometry)} that can be obtained from the function \code{allocate_reservoir_to_river()}}
}
\value{
the column \code{res_down} in the geospatial dataframe \code{reservoir_geometry_subset}
}
\description{
This function identifies which reservoirs are part of the river network and which reservoir do they drain into
}
