% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{route_reservoir_to_river}
\alias{route_reservoir_to_river}
\title{Routing reservoirs to the river network}
\usage{
route_reservoir_to_river(res_geom)
}
\arguments{
\item{res_geom}{is  a subset \code{data(reservoir_geometry)} that can be obtained from the function \code{allocate_reservoir_to_river()}}
}
\value{
completes the column \code{res_down} in the geospatial dataframe \code{reservoir_geometry}
}
\description{
This function identifies which reservoir is outside the river network and which river reach is closer and within the same catchment
}
