% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_topology.r
\name{select_disjoint_river}
\alias{select_disjoint_river}
\title{Select disjoint river network from set of river networks based on reach id}
\usage{
select_disjoint_river(reach_id, riv_all)
}
\arguments{
\item{reach_id}{an integer obtained from columne \code{UPLAND_SKM} of \code{data(river_geometry)}}

\item{riv_all}{a list of disjoint river networks obtained from \code{split_river_network()}.}
}
\value{
riv_i a disjoint river network
}
\description{
Select disjoint river network from set of river networks based on reach id
}
