% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{sink_from_graph}
\alias{sink_from_graph}
\title{Obtain sink from graph. In case there are disjoint graphs the sink is considered to be of greatest topological order}
\usage{
sink_from_graph(flow_graph)
}
\arguments{
\item{flow_graph}{is a tidygraph object ideally obtained from flow_graph_in_polygon}
}
\description{
Obtain sink from graph. In case there are disjoint graphs the sink is considered to be of greatest topological order
}
