% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{sort_1_strategic}
\alias{sort_1_strategic}
\title{Helper function for building topology of reservoirs}
\usage{
sort_1_strategic(strategic, riv_l)
}
\arguments{
\item{strategic}{is a group of \code{res_geom_list} from \code{build_reservoir_topology()}. It is obtained by grouping by \code{nearest river}}

\item{riv_l}{is the river linestring corresponding to the group \code{res_geom_list} being handled ie the \code{nearest river}}
}
\value{
strategic_df a dataframe with no geometry attributes with a \code{UP_CELLS} column filled where possible
}
\description{
Sorts strategic reservoirs whenever there is only one on the river network
}
