% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{sort_reservoirs}
\alias{sort_reservoirs}
\title{Helper function for building topology or reservoirs}
\usage{
sort_reservoirs(res_geom_i)
}
\arguments{
\item{res_geom_i}{is a subset of \code{data(reservoir_geometry)} group-split by \verb{nearest river}}
}
\value{
df a dataframe with sorting attributes
}
\description{
Sets sorting attributes to all reservoirs by assigning them to nearest strategic whenever there is one and running one of the helper functions below
}
