% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{split_river_network}
\alias{split_river_network}
\title{Split river network into disjoint graphs}
\usage{
split_river_network(riv)
}
\arguments{
\item{riv}{a sf dataframe with a topologicaly valid river network. Use for example the HydroSheds dataset.}
}
\value{
riv_n a sf dataframe with a topologicaly valid river network with a membership label for each disjoint graph.
}
\description{
Split river network into disjoint graphs
}
