% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{water_balance}
\alias{water_balance}
\title{Calculate the water balance and route downstream}
\usage{
water_balance(neigh, state)
}
\arguments{
\item{neigh}{is a dataframe describing the neighborhood upstream of the reservoir}

\item{state}{is a dataframe describing the current state of the reservoirs -- volume, incoming runoff, maximum volume and overspill}
}
\value{
state the state of the reservoir after water balance concluded
}
\description{
Calculate the water balance and route downstream
}
