% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{allocate_reservoir_to_river}
\alias{allocate_reservoir_to_river}
\title{Allocate each reservoir to nearest river reach within a given subbasin}
\usage{
allocate_reservoir_to_river(riv_i, reservoirs = reservoir_geometry)
}
\arguments{
\item{riv_i}{a subset of river reaches from \code{data(river_geometry)}}

\item{reservoirs}{a set of polygons or multipolygons of class \code{sf} to be matched with the river network. Defaults to the northeast Brazil reservoir dataset}
}
\value{
reservoir_geometry_subset subset of the reservoir data frame with the respective attributed river reach and distance to river reach
}
\description{
Allocate each reservoir to nearest river reach within a given subbasin
}
