% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{build_reservoir_topology}
\alias{build_reservoir_topology}
\title{Build reservoir topology starting from a river graph, river geometry and reservoir location}
\usage{
build_reservoir_topology(res_geom, riv_geom, riv_graph)
}
\arguments{
\item{res_geom}{is  a subset of \code{data(reservoir_geometry)} that can be obtained from the function \code{allocate_reservoir_to_river()}}

\item{riv_geom}{is a river geometry that can be created with \code{select_disjoint_river()}}

\item{riv_graph}{is a river graph created with \code{riv2graph()} based on riv_geom}
}
\value{
a the column \code{res_down} in the geospatial dataframe \code{res_geom}
}
\description{
This function builds the topology of the reservoir network. It starts by identifying which reservoirs
are located over the river network (strategic reservoirs) and builds their topology by filling the \code{res_down} column.
Then it looks at reservoirs outside the river network (in general smaller reservoirs and here called non-strategic)
and assigns them to a strategic reservoir (with \code{sf::st_nearest_feature()}) in case there is one in the river reach
or simply assigns the next downstream strategic reservoir as \code{res_down}
}
