% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_topology.r
\name{route_reservoir_along_river}
\alias{route_reservoir_along_river}
\title{Routing reservoirs across the river network}
\usage{
route_reservoir_along_river(res_geom, riv_geom = river_geometry, riv_graph)
}
\arguments{
\item{res_geom}{is  a subset of \code{data(reservoir_geometry)} that can be obtained from the function \code{allocate_reservoir_to_river()}}

\item{riv_geom}{is a river geometry possibly created with created with \code{select_disjoint_river()}}

\item{riv_graph}{is a river graph created with \code{riv2graph()} based on riv_geom}
}
\value{
the column \code{res_down} in the geospatial dataframe \code{res_geom}
}
\description{
This function identifies which reservoirs are part of the river network and which reservoir do they drain into
}
